+++ connections.c
@@ -1,0 +1,11 @@
+
+//STRUCT AREA
+struct connections {
+    Router * rs;
+    Connect * Next;
+};
+//END OF STRUCT AREA
+void printRouterAndTerminal(Router * r, Terminal * t) {
+    printRouters(r);
+    printTerminals(t);
+}
