+++ connections.h
@@ -1,0 +1,13 @@
+#ifndef CONNECTIONS_H
+#define CONNECTIONS_H
+typedef struct connections Connect;
+
+/*
+--printRouterAndTerminal--
+Precondition: Router and terminal are inicilized.
+Input: Router and terminal lists.
+Output: None.
+Postcondition: All routers and terminals are printed.
+*/
+void printRouterAndTerminal(Router * r, Terminal * t);
+#endif
