+++ readFile.c
@@ -1,8 +1,6 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
-#include "readFile.h"
-#include "net.h"
 //STRUCT AREA
 struct command {
   int fCode;   //Function code.
