+++ terminal.h
@@ -1,17 +1,10 @@
-#ifndef NET_H
-#define NET_H
-
-typedef struct router Router;
+#ifndef TERMINAL_H
+#define TERMINAL_H
 typedef struct terminal Terminal;
 
-/*
---registerRouter--
-Precondition: Router * r is not inicialized.
-Input: Not inicialized router, router name and router carrier name.
-Output: Updated router list.
-Postcondition: Router is inicialized and on the list.
-*/
-Router * registerRouter(Router * r, char * n, char * o);
+void disconnectRouter(Terminal * t);
+Terminal * findTerminalbyRouter(Terminal * t, char * rn);
+Terminal * findTerminal(Terminal * tlist, char * name);
 /*
 --registerTerminal--
 Precondition: Terminal * t is not inicialized.
@@ -20,6 +13,7 @@ Output: Updated terminal list.
 Postcondition: Terminsl is inicialized and on the list.
 */
 Terminal * registerTerminal(Terminal * t, char * n, char * l);
+
 /*
 --linkRouterToTerminal--
 Precondition: Router and terminal are not linked.
@@ -28,22 +22,7 @@ Output: None.
 Postcondition: Terminal and router are linked.
 */
 void linkRouterToTerminal(char * rname, Router * rlist, char * tname, Terminal * tlist);
-/*
---printRouterAndTerminal--
-Precondition: Router and terminal are inicilized.
-Input: Router and terminal lists.
-Output: None.
-Postcondition: All routers and terminals are printed.
-*/
-void printRouterAndTerminal(Router * r, Terminal * t);
-/*
---removeRouter--
-Precondition: Router and terminal are inicilized.
-Input: Router and terminal lists and router name.
-Output: List with removed router.
-Postcondition: Router is removed from the network.
-*/
-Router * removeRouter(Router * r, Terminal * ter, char * rn);
+
 /*
 --removeTerminal--
 Precondition: Router and terminal are inicilized.
@@ -52,6 +31,7 @@ Output: List with removed terminal.
 Postcondition: Terminal is removed from the network.
 */
 Terminal * removeTerminal(Router * r, Terminal * ter, char * tn);
+
 /*
 --unlinkTerminal--
 Precondition: Terminal is linked to a router.
@@ -60,4 +40,5 @@ Output: None.
 Postcondition: Terminal hasn't any link.
 */
 void unlinkTerminal(char * tname, Terminal * tlist);
+void printTerminals(Terminal * t);
 #endif
